Ugly fix to square the images input.  See commit message for more details.
The program now constrains the image to the smallest dimension.  So a 50x40 image is converted to 40x40 image.

A better option could be to allow the user to specify a coordinate.  This coordinate would be the top left corner of the area of interest.  And a square image would be extracted from that point.